refactor: regularize the internal config api
This wraps the server config module that is used as the repository for
config info as if it were a dict-like. This would enable eventually just
subbing in a dict (e.g. from yaml). In the short term it also makes the
module have better behavior when no server config is provided.