feat: don't drop corpses with Yred
This is to create the illusion that Yred always reaps when a corpse is
dropped. (Yred's reap chance can be improved with faith, so it's a
separate roll, but without faith they're both coinflips.)

For consistency, both the reaping brand effect and this Yred effect now
occur before dropping a hide. You cannot skin a zombie after its been
reanimated!